requirements.txt
torch>=1.10
torchvision>=0.11
Pillow
scikit-image
lpips
numpy
matplotlib
